WIP change move to integer distances, diagonals are integer too

+++ seawar/Move.cs
@@ -1,14 +1,14 @@
 ﻿namespace seawar
 {
    public class Move {
-      public Move(Vec vector, double distance, double speed) {
+      public Move(Vec vector, int distance, double speed) {
          Vector = vector;
          Distance = distance;
          Speed = speed;
       }
 
       public Vec Vector { get; }
-      public double Distance { get; }
+      public int Distance { get; }
       public double Speed { get; }
    }
 }
